make demo under the /swipe url extension
