+++ src/assets/data/people_data.ts
@@ -1,0 +1,7 @@
+export const sampleCards = [
+    { id: 1, name: 'Abe', num_reviews: 0, raw_total: 0 },
+    { id: 2, name: 'Ashley', num_reviews: 0, raw_total: 0 },
+    { id: 3, name: 'Denver', num_reviews: 0, raw_total: 0 },
+    { id: 4, name: 'Mario', num_reviews: 0, raw_total: 0 },
+    { id: 5, name: 'Joe', num_reviews: 0, raw_total: 0 }
+];
